Refactor NavLinks to Navbar - Colocate Avatar component logic
